LayoutTests/imported/w3c/web-platform-tests/keyboard-lock/idlharness.https.window-expected.txt


PASS idl_test setup
PASS idl_test validation
PASS Partial interface Navigator: original interface defined
PASS Partial interface Navigator: member names are unique
PASS Partial interface Navigator[2]: member names are unique
PASS Partial interface mixin NavigatorID: member names are unique
PASS Navigator includes NavigatorID: member names are unique
PASS Navigator includes NavigatorLanguage: member names are unique
PASS Navigator includes NavigatorOnLine: member names are unique
PASS Navigator includes NavigatorContentUtils: member names are unique
PASS Navigator includes NavigatorCookies: member names are unique
PASS Navigator includes NavigatorPlugins: member names are unique
PASS Navigator includes NavigatorConcurrentHardware: member names are unique
FAIL Keyboard interface: existence and properties of interface object assert_own_property: self does not have own property "Keyboard" expected property "Keyboard" missing
FAIL Keyboard interface object length assert_own_property: self does not have own property "Keyboard" expected property "Keyboard" missing
FAIL Keyboard interface object name assert_own_property: self does not have own property "Keyboard" expected property "Keyboard" missing
FAIL Keyboard interface: existence and properties of interface prototype object assert_own_property: self does not have own property "Keyboard" expected property "Keyboard" missing
FAIL Keyboard interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "Keyboard" expected property "Keyboard" missing
FAIL Keyboard interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "Keyboard" expected property "Keyboard" missing
FAIL Keyboard interface: operation lock(optional sequence<DOMString>) assert_own_property: self does not have own property "Keyboard" expected property "Keyboard" missing
FAIL Keyboard interface: operation unlock() assert_own_property: self does not have own property "Keyboard" expected property "Keyboard" missing
FAIL Keyboard must be primary interface of navigator.keyboard assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of navigator.keyboard assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Keyboard interface: navigator.keyboard must inherit property "lock(optional sequence<DOMString>)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Keyboard interface: calling lock(optional sequence<DOMString>) on navigator.keyboard with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Keyboard interface: navigator.keyboard must inherit property "unlock()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Navigator interface: attribute keyboard assert_true: The prototype object must have a property "keyboard" expected true got false
FAIL Navigator interface: navigator must inherit property "keyboard" with the proper type assert_inherits: property "keyboard" not found in prototype chain
